requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.20
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
whitenoise==6.4.0
